[api] - Refactoring
Moving the app/helpers/api_helper* files to app/controllers/api_controller/
as these are only used by the ApiController.
