+++ app/controllers/api_controller/results.rb
@@ -1,4 +1,4 @@
-module ApiHelper
+class ApiController
   module Results
     private
